requirements.txt
fastapi
uvicorn
openai-whisper
moviepy
pydub
pillow
ffmpeg-python
